[level 0] Title: 구슬을 나누는 경우의 수, Time: 0.02 ms, Memory: 75.9 MB -BaekjoonHub

+++ 프로그래머스/lv0/120840. 구슬을 나누는 경우의 수/구슬을 나누는 경우의 수.java
@@ -1,12 +1,12 @@
 class Solution {
-    public long solution(int balls, int share) {
-        long answer = 1;
+    public int solution(int balls, int share) {
+        double answer = 1;
         
         for(int i=share-1;i>=0;i--){
             answer *= balls-i;
             answer /= share-i;
         }
         
-        return answer;
+        return (int)answer;
     }
 }
